Slim chart semantic adapters onto core projection
