+++ crates/fricon/src/lib.rs
@@ -22,12 +22,14 @@ pub use self::{
         ColumnMeaning, ColumnMetadata, DatasetArray, DatasetDataType, DatasetEvent, DatasetId,
         DatasetInterpretation, DatasetListQuery, DatasetMetadata, DatasetReader, DatasetRecord,
         DatasetRow, DatasetScalar, DatasetSchema, DatasetSortBy, DatasetStatus, DatasetUpdate,
-        FixedStepTrace, InterpretationSource, PhysicalColumnOrdinal, ResolvedColumn,
-        ResolvedDuplicatePolicy, ResolvedIndexRealization, ResolvedLogicalIndexPoint,
-        ResolvedLogicalIndexReference, ResolvedPhysicalColumnReference, ResolvedScanAxis,
-        ResolvedScanAxisMode, ResolvedSemanticReference, ScalarArray, ScalarKind, ScanAxis,
-        ScanAxisMode, ScanAxisValue, ScanPlan, SelectOptions, SortDirection, TraceKind,
+        FixedStepTrace, InterpretationSource, PhysicalColumnOrdinal, ProjectedSemanticAxis,
+        ProjectedSemanticRoles, ProjectedSemanticSource, ResolvedColumn, ResolvedDuplicatePolicy,
+        ResolvedIndexRealization, ResolvedLogicalIndexPoint, ResolvedLogicalIndexReference,
+        ResolvedPhysicalColumnReference, ResolvedScanAxis, ResolvedScanAxisMode,
+        ResolvedSemanticReference, ScalarArray, ScalarKind, ScanAxis, ScanAxisMode, ScanAxisValue,
+        ScanPlan, SelectOptions, SemanticProjectionOptions, SortDirection, TraceKind,
         VariableStepTrace, VisibleColumnOrdinal, logical_index_id, physical_column_id,
+        project_semantic_source,
     },
     workspace::{WorkspaceError, WorkspaceRoot, get_log_dir},
 };
